[FIX] #23 UserWorkplace GET API
